Add a dummy presenter to test the mediaservice code.
