+++ Dockerfile.dummy_presenter
@@ -1,0 +1,12 @@
+FROM python:3.8.1
+
+WORKDIR /app
+
+COPY requirements_production.txt .
+COPY requirements_development.txt .
+RUN pip install -r requirements_development.txt
+
+ENV FLASK_APP ./dummy_presenter/dummy_presenter.py
+ENV FLASK_ENV development
+
+CMD ["flask", "run", "--host", "0.0.0.0", "--port", "9003"]
