+++ Makefile
@@ -4,7 +4,10 @@ build-dev:
 build-tests:
 	docker build . -f Dockerfile.tests --tag openslides-media-tests
 
-run-tests: | build-dev build-tests
+build-dummy-presenter:
+	docker build . -f Dockerfile.dummy_presenter --tag openslides-media-dummy-presenter
+
+run-tests: | build-dev build-tests build-dummy-presenter
 	docker-compose -f docker-compose.test.yml up -d
 	docker-compose -f docker-compose.test.yml exec tests wait-for-it "media:9006"
 	docker-compose -f docker-compose.test.yml exec tests pytest
